deployment/artifact/requirements.txt
Click==7.0
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
Werkzeug==0.15.4
